Consolidating controller feature into another controller
